LPS-112887 We want to remove any dependency from headless API in the service, so move everything to exporter package since we are going to have an importer package too
