+++ modules/apps/layout/layout-page-template-admin-web/src/main/java/com/liferay/layout/page/template/admin/web/internal/headless/delivery/dto/v1_0/structure/exporter/LayoutStructureItemExporter.java
@@ -12,36 +12,20 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure;
+package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure.exporter;
 
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
 import com.liferay.layout.util.structure.LayoutStructureItem;
-import com.liferay.layout.util.structure.RootLayoutStructureItem;
-
-import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Jürgen Kappler
  */
-@Component(service = LayoutStructureItemExporter.class)
-public class RootLayoutStructureItemExporter
-	implements LayoutStructureItemExporter {
+public interface LayoutStructureItemExporter {
 
-	@Override
-	public String getClassName() {
-		return RootLayoutStructureItem.class.getName();
-	}
+	public String getClassName();
 
-	@Override
 	public PageElement getPageElement(
 		long groupId, LayoutStructureItem layoutStructureItem,
-		boolean saveInlineContent, boolean saveMappingConfiguration) {
-
-		return new PageElement() {
-			{
-				type = PageElement.Type.ROOT;
-			}
-		};
-	}
+		boolean saveInlineContent, boolean saveMappingConfiguration);
 
 }
